Code refactoring to allow for expansion
This code is being modified for use in research involving flux measurements with open path line concentrations.  The calculations have not changed much; however, the code was reorganized to reduce creep and make the code more modular.  More modular should make new features easier to add. Some new features have been added including:

(1) Warning messages are formatted identically so it is clear the windrose function is generating the warning and it is not coming from outside the windrose function.
(2) Guards have been placed to check if the incoming data is in the correct form.  This may alleviate issue #103.  In the original implementation, if the incoming data is formatted wrong it is hard to tell what is wrong because errors may be anything generated anywhere.   
(3) More warnings are included so the user knows about situations like datapoints falling outside of the ws.int and/or breaks specifications.
(4) A maximum can now be specified for the interval rings.  This is to prevent the graph from displaying unrealistic frequency lines greater than 100%
windRose(mydata, grid.line = list(max = 100))
(5) The width of grid lines can be adjusted
windRose(mydata, grid.line = list(lwd = 2))
(6) For big plots a finer circle can be made by doubling the resolution factor
WindRose(mydata, grid.line = list(res = 2))
(7) An anticipated direction arrow can be added.  This is useful for determining if the wind blew in an anticipated direction.
windRose(mydata, grid.line = list(gridArrow = 180))
(8) The breaks are calculated differently.  If nothing is given pretty values are used.  If a vector is given then the vector is used.  If a vector is given, but not all the points are included a warning is issued before adding a break.
(9) dig.lab default was changed to 0... Not sure if this was for the best, but it is easy to change back.  It may be worth it to accept input as a named list so specific variables can be controlled.

Internal Changes to the code:
(1) mydata is no longer treated as a mutable variable.  Changes are saved in the statdata variable instead.
(2) windspeed and pollution are better separated. If a variable says wind speed then it is more likely windspeed.
(3) As best as possible limited the characters per line to 80 so it displays well on most screens.
(4) As best as possible removed dots (.) from name to avoid confusion with the dot operator.  lower camel case was used instead.
(5) Avoided repetition of code as much as possible.
(6) Used functions to delineate tasks such as drawing the individual elements.  This should make the code much simpler to upgrade.
(7) the name 'results' has been changed to 'pltData' because with results it is confusing about which results. 
(8) Angle360 function automatically coerces angles between (0 - 360]
(9) DegtoRad performs the degrees to radians calculation
(8)  One computation was confusing...  

First the wind direction angle is set into intervals
<code>
line 533 -  
mydata[[wd]] <- angle * ceiling(mydata[[wd]] / angle - 0.5)
</code>

Then the average wind direction is calculated from angles set to intervals
<code>
line 610 -  ## calculate mean wd - useful for cases comparing two met data sets
-      u <- mean(sin(2 * pi * mydata[[wd]] / 360), na.rm = TRUE)
-      v <- mean(cos(2 * pi * mydata[[wd]] / 360), na.rm = TRUE)
-      mean.wd <- atan2(u, v) * 360 / 2 / pi
</code>

The proposed version works differently averages wind direction from non-interval data.

(9) The concept of an apothem is introduced to make understanding lengths in the plot simple.  The apothem of a square goes from the center to the mid-point of one of the sides.


In progress:
(1) Plotting a point with the average wind speed and direction directly on the wind rose.  This will be used as a means to determine the quality of an average wind speed and direction.  This feature is already included, but the average wind speed is in the wrong spot.