Adding the start of a vignette
Vignettes of Openair functions seem like a very good idea.  Not only for educational purposes and for those new to R, but also for testing purposes.  Each example can double as a unit test.  When the vignette is knitted, it is very easy to tell if and when the code breaks.

Modifications to the R code can be tested against the vignette specific to the function.