removing unneeded repitition
Functionality is identical except with 30 fewer lines.  This represents 17 edits to remove unneeded repetition from the code.  Instead of attempting a major refactoring, a series of minor refactorings may be more effective.   My interest in working on this file is primarily to gain practical experience with the Github platform as well as learning how to push the capability of this software.